- Move NationPreMergeEvent, NationMergeEvent, PreNewNationEvent to nation subpackage.
